Moved SolidGaugeAxis to Axis folder
